+++ src/taiko/gradual_difficulty.rs
@@ -390,7 +390,7 @@ mod tests {
     fn iter_end_eq_regular() {
         let map = Beatmap::from_path("./maps/1028484.osu").expect("failed to parse map");
         let mods = 64;
-        let regular = crate::taiko::stars(&map, mods, None);
+        let regular = crate::TaikoStars::new(&map).mods(mods).calculate();
 
         let iter_end = TaikoGradualDifficultyAttributes::new(&map, mods)
             .last()